requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.2
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
